// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  files: typeof files;
  flipbooks: typeof flipbooks;
  "lib/session": typeof lib_session;
  posterGenerations: typeof posterGenerations;
  posterInternal: typeof posterInternal;
  posters: typeof posters;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;